Добавляет список комментариев
